Move krate::new to krate::publish::publish
